OrdinaryDiffEqLowOrderRK

If OrdinaryDiffEqTsit5 is not working well for your non-stiff problem at default and higher tolerance, it can be worthwhile to explore the options in this package. In particular, when more robust error control is required, BS5 is a good choice. If at moderate tolerances and the interpolation error is very important, consider the OwrenZen5 method. For fast solving at higher tolerances, we recommend BS3, or OwrenZen3if the interpolation error is important.

Installation

To be able to access the solvers in OrdinaryDiffEqLowOrderRK, you must first install them use the Julia package manager:

using Pkg
Pkg.add("OrdinaryDiffEqLowOrderRK")

This will only install the solvers listed at the bottom of this page. If you want to explore other solvers for your problem, you will need to install some of the other libraries listed in the navigation bar on the left.

Example usage

using OrdinaryDiffEqLowOrderRK

function lorenz!(du, u, p, t)
    du[1] = 10.0 * (u[2] - u[1])
    du[2] = u[1] * (28.0 - u[3]) - u[2]
    du[3] = u[1] * u[2] - (8 / 3) * u[3]
end
u0 = [1.0; 0.0; 0.0]
tspan = (0.0, 100.0)
prob = ODEProblem(lorenz!, u0, tspan)
sol = solve(prob, BS3())

Full list of solvers

OrdinaryDiffEqLowOrderRK.Euler — Type

Euler()

Explicit Runge-Kutta Method. The canonical forward Euler method. Fixed timestep only.

Keyword Arguments

References

E. Hairer, S.P. Norsett, G. Wanner, (1993) Solving Ordinary Differential Equations I. Nonstiff Problems. 2nd Edition. Springer Series in Computational Mathematics, Springer-Verlag.

OrdinaryDiffEqLowOrderRK.Heun — Type

Heun(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
       step_limiter! = OrdinaryDiffEq.trivial_limiter!,
       thread = OrdinaryDiffEq.False())

Explicit Runge-Kutta Method. The second order Heun's method. Uses embedded Euler method for adaptivity.

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False()) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

References

E. Hairer, S.P. Norsett, G. Wanner, (1993) Solving Ordinary Differential Equations I. Nonstiff Problems. 2nd Edition. Springer Series in Computational Mathematics, Springer-Verlag.

OrdinaryDiffEqLowOrderRK.Ralston — Type

Ralston(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
          step_limiter! = OrdinaryDiffEq.trivial_limiter!,
          thread = OrdinaryDiffEq.False())

Explicit Runge-Kutta Method. The optimized second order midpoint method. Uses embedded Euler method for adaptivity.

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False()) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

References

E. Hairer, S.P. Norsett, G. Wanner, (1993) Solving Ordinary Differential Equations I. Nonstiff Problems. 2nd Edition. Springer Series in Computational Mathematics, Springer-Verlag.

OrdinaryDiffEqLowOrderRK.Midpoint — Type

Midpoint(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
           step_limiter! = OrdinaryDiffEq.trivial_limiter!,
           thread = OrdinaryDiffEq.False())

Explicit Runge-Kutta Method. The second order midpoint method. Uses embedded Euler method for adaptivity.

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False()) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

References

E. Hairer, S.P. Norsett, G. Wanner, (1993) Solving Ordinary Differential Equations I. Nonstiff Problems. 2nd Edition. Springer Series in Computational Mathematics, Springer-Verlag.

OrdinaryDiffEqLowOrderRK.RK4 — Type

RK4(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
      step_limiter! = OrdinaryDiffEq.trivial_limiter!,
      thread = OrdinaryDiffEq.False())

Explicit Runge-Kutta Method. The canonical Runge-Kutta Order 4 method. Uses a defect control for adaptive stepping using maximum error over the whole interval.

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False()) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

References

@article{shampine2005solving, title={Solving ODEs and DDEs with residual control}, author={Shampine, LF}, journal={Applied Numerical Mathematics}, volume={52}, number={1}, pages={113–127}, year={2005}, publisher={Elsevier} }

OrdinaryDiffEqLowOrderRK.BS3 — Type

BS3(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
      step_limiter! = OrdinaryDiffEq.trivial_limiter!,
      thread = OrdinaryDiffEq.False())

Explicit Runge-Kutta Method. A third-order, four-stage FSAL method with embedded error estimator of Bogacki and Shampine.

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False()) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

References

@article{bogacki19893, title={A 3 (2) pair of Runge-Kutta formulas}, author={Bogacki, Przemyslaw and Shampine, Lawrence F}, journal={Applied Mathematics Letters}, volume={2}, number={4}, pages={321–325}, year={1989}, publisher={Elsevier} }

OrdinaryDiffEqLowOrderRK.OwrenZen3 — Type

OwrenZen3(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
            step_limiter! = OrdinaryDiffEq.trivial_limiter!,
            thread = OrdinaryDiffEq.False())

Explicit Runge-Kutta Method. Owren-Zennaro optimized interpolation 3/2 method (free 3rd order interpolant).

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False()) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

References

@article{owren1992derivation, title={Derivation of efficient, continuous, explicit Runge–Kutta methods}, author={Owren, Brynjulf and Zennaro, Marino}, journal={SIAM journal on scientific and statistical computing}, volume={13}, number={6}, pages={1488–1501}, year={1992}, publisher={SIAM} }

OrdinaryDiffEqLowOrderRK.OwrenZen4 — Type

OwrenZen4(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
            step_limiter! = OrdinaryDiffEq.trivial_limiter!,
            thread = OrdinaryDiffEq.False())

Explicit Runge-Kutta Method. Owren-Zennaro optimized interpolation 4/3 method (free 4th order interpolant).

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False()) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

References

@article{owren1992derivation, title={Derivation of efficient, continuous, explicit Runge–Kutta methods}, author={Owren, Brynjulf and Zennaro, Marino}, journal={SIAM journal on scientific and statistical computing}, volume={13}, number={6}, pages={1488–1501}, year={1992}, publisher={SIAM} }

OrdinaryDiffEqLowOrderRK.OwrenZen5 — Type

OwrenZen5(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
            step_limiter! = OrdinaryDiffEq.trivial_limiter!,
            thread = OrdinaryDiffEq.False())

Explicit Runge-Kutta Method. Owren-Zennaro optimized interpolation 5/4 method (free 5th order interpolant).

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False()) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

References

@article{owren1992derivation, title={Derivation of efficient, continuous, explicit Runge–Kutta methods}, author={Owren, Brynjulf and Zennaro, Marino}, journal={SIAM journal on scientific and statistical computing}, volume={13}, number={6}, pages={1488–1501}, year={1992}, publisher={SIAM} }

OrdinaryDiffEqLowOrderRK.BS5 — Type

BS5(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
      step_limiter! = OrdinaryDiffEq.trivial_limiter!,
      thread = OrdinaryDiffEq.False(),
      lazy = true)

Explicit Runge-Kutta Method. Bogacki-Shampine 5/4 Runge-Kutta method. (lazy 5th order interpolant).

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False()) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

lazy: determines if the lazy interpolant is used.

References

@article{bogacki1996efficient, title={An efficient runge-kutta (4, 5) pair}, author={Bogacki, P and Shampine, Lawrence F}, journal={Computers \& Mathematics with Applications}, volume={32}, number={6}, pages={15–28}, year={1996}, publisher={Elsevier} }

OrdinaryDiffEqLowOrderRK.DP5 — Type

DP5(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
      step_limiter! = OrdinaryDiffEq.trivial_limiter!,
      thread = OrdinaryDiffEq.False())

Explicit Runge-Kutta Method. Dormand-Prince's 5/4 Runge-Kutta method. (free 4th order interpolant).

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False()) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

References

@article{dormand1980family, title={A family of embedded Runge-Kutta formulae}, author={Dormand, John R and Prince, Peter J}, journal={Journal of computational and applied mathematics}, volume={6}, number={1}, pages={19–26}, year={1980}, publisher={Elsevier} }

OrdinaryDiffEqLowOrderRK.Anas5 — Type

Anas5(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
        step_limiter! = OrdinaryDiffEq.trivial_limiter!,
        thread = OrdinaryDiffEq.False(),
        w = 1)

Explicit Runge-Kutta Method. 4th order Runge-Kutta method designed for periodic problems.

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False()) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

w: a periodicity estimate, which when accurate the method becomes 5th order

(and is otherwise 4th order with less error for better estimates).

References

@article{anastassi2005optimized, title={An optimized Runge–Kutta method for the solution of orbital problems}, author={Anastassi, ZA and Simos, TE}, journal={Journal of Computational and Applied Mathematics}, volume={175}, number={1}, pages={1–9}, year={2005}, publisher={Elsevier}}

OrdinaryDiffEqLowOrderRK.RKO65 — Type

RKO65(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
        step_limiter! = OrdinaryDiffEq.trivial_limiter!,
        thread = OrdinaryDiffEq.False())

Explicit Runge-Kutta Method. 5th order method.

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False()) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

References

Tsitouras, Ch. "Explicit Runge–Kutta methods for starting integration of Lane–Emden problem." Applied Mathematics and Computation 354 (2019): 353-364. doi: https://doi.org/10.1016/j.amc.2019.02.047

OrdinaryDiffEqLowOrderRK.FRK65 — Type

FRK65(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
        step_limiter! = OrdinaryDiffEq.trivial_limiter!,
        thread = OrdinaryDiffEq.False(),
        omega = 0.0)

Explicit Runge-Kutta Method. Zero Dissipation Runge-Kutta of 6th order.

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False()) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

omega: a periodicity phase estimate,

when accurate this method results in zero numerical dissipation.

References

@article{medvedev2018fitted, title={Fitted modifications of Runge-Kutta pairs of orders 6 (5)}, author={Medvedev, Maxim A and Simos, TE and Tsitouras, Ch}, journal={Mathematical Methods in the Applied Sciences}, volume={41}, number={16}, pages={6184–6194}, year={2018}, publisher={Wiley Online Library}}

OrdinaryDiffEqLowOrderRK.RKM — Type

RKM(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
      step_limiter! = OrdinaryDiffEq.trivial_limiter!,
      thread = OrdinaryDiffEq.False())

Explicit Runge-Kutta Method. Method designed to have good stability properties when applied to pseudospectral discretizations of hyperbolic partial differential equaitons.

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False()) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

References

@article{mead1999optimal, title={Optimal Runge–Kutta methods for first order pseudospectral operators}, author={Mead, JL and Renaut, RA}, journal={Journal of Computational Physics}, volume={152}, number={1}, pages={404–419}, year={1999}, publisher={Elsevier} }

OrdinaryDiffEqLowOrderRK.MSRK5 — Type

MSRK5(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
        step_limiter! = OrdinaryDiffEq.trivial_limiter!,
        thread = OrdinaryDiffEq.False())

Explicit Runge-Kutta Method. 5th order method.

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False()) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

References

Misha Stepanov - https://arxiv.org/pdf/2202.08443.pdf : Figure 3.

OrdinaryDiffEqLowOrderRK.MSRK6 — Type

MSRK6(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
        step_limiter! = OrdinaryDiffEq.trivial_limiter!,
        thread = OrdinaryDiffEq.False())

Explicit Runge-Kutta Method. 6th order method.

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False()) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

References

Misha Stepanov - https://arxiv.org/pdf/2202.08443.pdf : Table4

OrdinaryDiffEqLowOrderRK.PSRK4p7q6 — Type

PSRK4p7q6(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
            step_limiter! = OrdinaryDiffEq.trivial_limiter!,
            thread = OrdinaryDiffEq.False())

Explicit Runge-Kutta Method. 6-stage Pseudo-Symplectic method.

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False()) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

References

@article{Aubry1998, author = {A. Aubry and P. Chartier}, journal = {BIT Numer. Math.}, title = {Pseudo-symplectic {R}unge-{K}utta methods}, volume = {38}, PAGES = {439-461}, year = {1998}, }, @article{Capuano2017, title = {Explicit {R}unge–{K}utta schemes for incompressible flow with improved energy-conservation properties}, journal = {J. Comput. Phys.}, volume = {328}, pages = {86-94}, year = {2017}, issn = {0021-9991}, doi = {https://doi.org/10.1016/j.jcp.2016.10.040}, author = {F. Capuano and G. Coppola and L. Rández and L. {de Luca}},}

OrdinaryDiffEqLowOrderRK.PSRK3p5q4 — Type

PSRK3p5q4(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
            step_limiter! = OrdinaryDiffEq.trivial_limiter!,
            thread = OrdinaryDiffEq.False())

Explicit Runge-Kutta Method. 4-stage Pseudo-Symplectic method.

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False()) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

References

@article{Aubry1998, author = {A. Aubry and P. Chartier}, journal = {BIT Numer. Math.}, title = {Pseudo-symplectic {R}unge-{K}utta methods}, year = {1998}, }, @article{Capuano2017, title = {Explicit {R}unge–{K}utta schemes for incompressible flow with improved energy-conservation properties}, journal = {J. Comput. Phys.}, year = {2017}, author = {F. Capuano and G. Coppola and L. Rández and L. {de Luca}},}

OrdinaryDiffEqLowOrderRK.PSRK3p6q5 — Type

PSRK3p6q5(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
            step_limiter! = OrdinaryDiffEq.trivial_limiter!,
            thread = OrdinaryDiffEq.False())

Explicit Runge-Kutta Method. 5-stage Pseudo-Symplectic method.

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False()) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

References

@article{Aubry1998, author = {A. Aubry and P. Chartier}, journal = {BIT Numer. Math.}, title = {Pseudo-symplectic {R}unge-{K}utta methods}, year = {1998}, }, @article{Capuano2017, title = {Explicit {R}unge–{K}utta schemes for incompressible flow with improved energy-conservation properties}, journal = {J. Comput. Phys.}, year = {2017}, author = {F. Capuano and G. Coppola and L. Rández and L. {de Luca}},}

OrdinaryDiffEqLowOrderRK.Stepanov5 — Type

Stepanov5(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
            step_limiter! = OrdinaryDiffEq.trivial_limiter!,
            thread = OrdinaryDiffEq.False())

Explicit Runge-Kutta Method. 5th order method.

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False()) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

References

@article{Stepanov2021Embedded5, title={Embedded (4, 5) pairs of explicit 7-stage Runge–Kutta methods with FSAL property}, author={Misha Stepanov}, journal={Calcolo}, year={2021}, volume={59} }

OrdinaryDiffEqLowOrderRK.SIR54 — Type

SIR54(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
        step_limiter! = OrdinaryDiffEq.trivial_limiter!,
        thread = OrdinaryDiffEq.False())

Explicit Runge-Kutta Method. 5th order method suited for SIR-type epidemic models.

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False()) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

References

@article{Kovalnogov2020RungeKuttaPS, title={Runge–Kutta pairs suited for SIR‐type epidemic models}, author={Vladislav N. Kovalnogov and Theodore E. Simos and Ch. Tsitouras}, journal={Mathematical Methods in the Applied Sciences}, year={2020}, volume={44}, pages={5210 - 5216} }

OrdinaryDiffEqLowOrderRK.Alshina2 — Type

Alshina2(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
           step_limiter! = OrdinaryDiffEq.trivial_limiter!,
           thread = OrdinaryDiffEq.False())

Explicit Runge-Kutta Method. 2nd order, 2-stage Method with optimal parameters.

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False()) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

References

@article{Alshina2008, doi = {10.1134/s0965542508030068}, url = {https://doi.org/10.1134/s0965542508030068}, year = {2008}, month = mar, publisher = {Pleiades Publishing Ltd}, volume = {48}, number = {3}, pages = {395–405}, author = {E. A. Alshina and E. M. Zaks and N. N. Kalitkin}, title = {Optimal first- to sixth-order accurate Runge-Kutta schemes}, journal = {Computational Mathematics and Mathematical Physics} }

OrdinaryDiffEqLowOrderRK.Alshina3 — Type

Alshina3(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
           step_limiter! = OrdinaryDiffEq.trivial_limiter!,
           thread = OrdinaryDiffEq.False())

Explicit Runge-Kutta Method. 3rd order, 3-stage Method with optimal parameters.

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False()) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

References

@article{Alshina2008, doi = {10.1134/s0965542508030068}, url = {https://doi.org/10.1134/s0965542508030068}, year = {2008}, month = mar, publisher = {Pleiades Publishing Ltd}, volume = {48}, number = {3}, pages = {395–405}, author = {E. A. Alshina and E. M. Zaks and N. N. Kalitkin}, title = {Optimal first- to sixth-order accurate Runge-Kutta schemes}, journal = {Computational Mathematics and Mathematical Physics} }

OrdinaryDiffEqLowOrderRK.Alshina6 — Type

Alshina6(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
           step_limiter! = OrdinaryDiffEq.trivial_limiter!,
           thread = OrdinaryDiffEq.False())

Explicit Runge-Kutta Method. 6th order, 7-stage Method with optimal parameters.

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False()) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

References

@article{Alshina2008, doi = {10.1134/s0965542508030068}, url = {https://doi.org/10.1134/s0965542508030068}, year = {2008}, month = mar, publisher = {Pleiades Publishing Ltd}, volume = {48}, number = {3}, pages = {395–405}, author = {E. A. Alshina and E. M. Zaks and N. N. Kalitkin}, title = {Optimal first- to sixth-order accurate Runge-Kutta schemes}, journal = {Computational Mathematics and Mathematical Physics} }