OrdinaryDiffEqQPRK

Explicit solvers optimized for a certain number of parallel calls of the system of ordinary differential equations f. Particularly good at low tolerances, when using quad-precision arithmetic, Float128.

Installation

To be able to access the solvers in OrdinaryDiffEqQPRK, you must first install them use the Julia package manager:

using Pkg
Pkg.add("OrdinaryDiffEqQPRK")

This will only install the solvers listed at the bottom of this page. If you want to explore other solvers for your problem, you will need to install some of the other libraries listed in the navigation bar on the left.

Example usage

using OrdinaryDiffEqQPRK

function lorenz!(du, u, p, t)
    du[1] = 10.0 * (u[2] - u[1])
    du[2] = u[1] * (28.0 - u[3]) - u[2]
    du[3] = u[1] * u[2] - (8 / 3) * u[3]
end
u0 = [1.0; 0.0; 0.0]
tspan = (0.0, 100.0)
prob = ODEProblem(lorenz!, u0, tspan)
sol = solve(prob, QPRK98())

Full list of solvers

OrdinaryDiffEqQPRK.QPRK98 — Type

QPRK98(; stage_limiter! = OrdinaryDiffEq.trivial_limiter!,
         step_limiter! = OrdinaryDiffEq.trivial_limiter!,
         thread = OrdinaryDiffEq.False())

Explicit Runge-Kutta Method. Runge–Kutta pairs of orders 9(8) for use in quadruple precision computations

Keyword Arguments

stage_limiter!: function of the form limiter!(u, integrator, p, t)

step_limiter!: function of the form limiter!(u, integrator, p, t)

thread: determines whether internal broadcasting on appropriate CPU arrays should be serial (thread = OrdinaryDiffEq.False()) or use multiple threads (thread = OrdinaryDiffEq.True()) when Julia is started with multiple threads.

References

Kovalnogov VN, Fedorov RV, Karpukhina TV, Simos TE, Tsitouras C. Runge–Kutta pairs of orders 9 (8) for use in quadruple precision computations. Numerical Algorithms, 2023. doi: https://doi.org/10.1007/s11075-023-01632-8