OrdinaryDiffEqIMEXMultistep

Solvers if your system of ordinary differential equations can be split up into the sum of a stiff and non-stiff part. These are "standard" low order IMEX methods common in many naive partial differential equation applications.

Installation

To be able to access the solvers in OrdinaryDiffEqIMEXMultistep, you must first install them use the Julia package manager:

using Pkg
Pkg.add("OrdinaryDiffEqIMEXMultistep")

This will only install the solvers listed at the bottom of this page. If you want to explore other solvers for your problem, you will need to install some of the other libraries listed in the navigation bar on the left.

Example usage

using OrdinaryDiffEqIMEXMultistep
f1 = (u, p, t) -> 2u
f2 = (u, p, t) -> 2u
u0 = 1.0
tspan = (0.0, 1.0)
prob = SplitODEProblem(f1, f2, u0, tspan)
sol = solve(prob, CNAB2(), dt = 1 / 10)

Full list of solvers

OrdinaryDiffEqIMEXMultistep.CNAB2 — Type

CNAB2()

IMEX Multistep method. Crank-Nicholson Adams-Bashforth 2.

Keyword Arguments

References

@article{jorgenson2014unconditional, title={Unconditional stability of a Crank-Nicolson Adams-Bashforth 2 numerical method}, author={JORGENSON, ANDREW D}, journal={A (A- C)}, volume={1}, number={2}, pages={1}, year={2014}} @article{he2010numerical, title={Numerical implementation of the Crank–Nicolson/Adams–Bashforth scheme for the time-dependent Navier–Stokes equations}, author={He, Yinnian and Li, Jian}, journal={International journal for numerical methods in fluids}, volume={62}, number={6}, pages={647–659}, year={2010}, publisher={Wiley Online Library}}

OrdinaryDiffEqIMEXMultistep.CNLF2 — Type

CNLF2()

IMEX Multistep method. Crank-Nicholson Leapfrong 2.

Keyword Arguments

References

@article{han2020second, title={A second order, linear, unconditionally stable, Crank–Nicolson–Leapfrog scheme for phase field models of two-phase incompressible flows}, author={Han, Daozhi and Jiang, Nan}, journal={Applied Mathematics Letters}, volume={108}, pages={106521}, year={2020}, publisher={Elsevier}} @article{jiang2015crank, title={A Crank–Nicolson Leapfrog stabilization: Unconditional stability and two applications}, author={Jiang, Nan and Kubacki, Michaela and Layton, William and Moraiti, Marina and Tran, Hoang}, journal={Journal of Computational and Applied Mathematics}, volume={281}, pages={263–276}, year={2015}, publisher={Elsevier}}