OrdinaryDiffEqNordsieck

The Nordsieck form is an alternative representation of multistep methods which, instead of representing and saving past step values in a history vector, it uses a derivative list (like a Taylor expansion) for the computation of the next point. The Nordsieck form was pioneered by early implementations of BDF methods such LSODE, VODE, and finally CVODE. It can have some advantages in terms of restartability as the full Nordsieck vector can be instantiated given only the information of f and its derivatives after discontinuities, but the higher derivative representations can also introduce numerical instabilities of their own.

The Nordsieck implementations here are considered experimental implementations of the LSODE non-fixed leading coefficient form and are generally considered inferior to the fixed-leading history-based BDF implementation of FBDF, and thus for all standard usage we recommend FBDF. However, this algorithm is kept for experimental research and development purposes with the possibility of one day becoming a more discontinuity-aware BDF implementation.

Installation

To be able to access the solvers in OrdinaryDiffEqNordsieck, you must first install them use the Julia package manager:

using Pkg
Pkg.add("OrdinaryDiffEqNordsieck")

This will only install the solvers listed at the bottom of this page. If you want to explore other solvers for your problem, you will need to install some of the other libraries listed in the navigation bar on the left.

Example usage

using OrdinaryDiffEqNordsieck

function lorenz!(du, u, p, t)
    du[1] = 10.0 * (u[2] - u[1])
    du[2] = u[1] * (28.0 - u[3]) - u[2]
    du[3] = u[1] * u[2] - (8 / 3) * u[3]
end
u0 = [1.0; 0.0; 0.0]
tspan = (0.0, 100.0)
prob = ODEProblem(lorenz!, u0, tspan)
sol = solve(prob, AN5())

Full list of solvers

OrdinaryDiffEqNordsieck.AN5 — Type

AN5()

Adaptive step size Adams explicit Method An adaptive 5th order fixed-leading coefficient Adams method in Nordsieck form.

Experimental

AN5 is experimental, the solver VCABM is generally preferred.

Keyword Arguments

References

OrdinaryDiffEqNordsieck.JVODE — Type

Experimental

JVODE is experimental, the solver VCABM is generally preferred.

OrdinaryDiffEqNordsieck.JVODE_Adams — Function

Experimental

JVODE is experimental, the solver VCABM is generally preferred.

OrdinaryDiffEqNordsieck.JVODE_BDF — Function

Experimental

JVODE is experimental, the solver FBDF is generally preferred.