Explicit Extrapolation Methods

OrdinaryDiffEq.OrdinaryDiffEqExtrapolation.AitkenNeville — Type

AitkenNeville: Parallelized Explicit Extrapolation Method Euler extrapolation using Aitken-Neville with the Romberg Sequence.

OrdinaryDiffEq.OrdinaryDiffEqExtrapolation.ExtrapolationMidpointDeuflhard — Type

ExtrapolationMidpointDeuflhard: Parallelized Explicit Extrapolation Method Midpoint extrapolation using Barycentric coordinates

OrdinaryDiffEq.OrdinaryDiffEqExtrapolation.ExtrapolationMidpointHairerWanner — Type

ExtrapolationMidpointHairerWanner: Parallelized Explicit Extrapolation Method Midpoint extrapolation using Barycentric coordinates, following Hairer's ODEX in the adaptivity behavior.