OrdinaryDiffEqExponentialRK

Methods for semi-linear differential equations.

Installation

To be able to access the solvers in OrdinaryDiffEqLinear, you must first install them use the Julia package manager:

using Pkg
Pkg.add("OrdinaryDiffEqExponentialRK")

This will only install the solvers listed at the bottom of this page. If you want to explore other solvers for your problem, you will need to install some of the other libraries listed in the navigation bar on the left.

Example usage

Installation

To be able to access the solvers in OrdinaryDiffEqExponentialRK, you must first install them use the Julia package manager:

using Pkg
Pkg.add("OrdinaryDiffEqExponentialRK")

This will only install the solvers listed at the bottom of this page. If you want to explore other solvers for your problem, you will need to install some of the other libraries listed in the navigation bar on the left.

Example usage

using OrdinaryDiffEqExponentialRK

function lorenz!(du, u, p, t)
    du[1] = 10.0 * (u[2] - u[1])
    du[2] = u[1] * (28.0 - u[3]) - u[2]
    du[3] = u[1] * u[2] - (8 / 3) * u[3]
end
u0 = [1.0; 0.0; 0.0]
tspan = (0.0, 100.0)
prob = ODEProblem(lorenz!, u0, tspan)
sol = solve(prob, EPIRK5s3())

Full list of solvers

OrdinaryDiffEqExponentialRK.LawsonEuler — Type

LawsonEuler(; krylov = false,
              m = 30,
              iop = 0)

Semilinear ODE solver First order exponential Euler scheme.

Keyword Arguments

krylov: Determines whether Krylov approximation or operator caching is used, the latter only available for semilinear problems. krylov=true is much faster for larger systems and is thus recommended whenever there are >100 ODEs.

m: Controls the size of Krylov subspace.

iop: If not zero, determines the length of the incomplete orthogonalization procedure (IOP). Note that if the linear operator/Jacobian is hermitian, then the Lanczos algorithm will always be used and the IOP setting is ignored.

References

Hochbruck, Marlis, and Alexander Ostermann. “Exponential Integrators.” Acta Numerica 19 (2010): 209–286. doi:10.1017/S0962492910000048.

OrdinaryDiffEqExponentialRK.NorsettEuler — Type

NorsettEuler(; krylov = false,
               m = 30,
               iop = 0)

Semilinear ODE solver First order exponential-RK scheme. Alias: ETD1

Keyword Arguments

krylov: Determines whether Krylov approximation or operator caching is used, the latter only available for semilinear problems. krylov=true is much faster for larger systems and is thus recommended whenever there are >100 ODEs.

m: Controls the size of Krylov subspace.

iop: If not zero, determines the length of the incomplete orthogonalization procedure (IOP). Note that if the linear operator/Jacobian is hermitian, then the Lanczos algorithm will always be used and the IOP setting is ignored.

References

Hochbruck, Marlis, and Alexander Ostermann. “Exponential Integrators.” Acta Numerica 19 (2010): 209–286. doi:10.1017/S0962492910000048.

OrdinaryDiffEqExponentialRK.ETD2 — Type

ETD2: Exponential Runge-Kutta Method Second order Exponential Time Differencing method (in development).

OrdinaryDiffEqExponentialRK.ETDRK2 — Type

ETDRK2(; krylov = false,
         m = 30,
         iop = 0)

Semilinear ODE solver 2nd order exponential-RK scheme.

Keyword Arguments

krylov: Determines whether Krylov approximation or operator caching is used, the latter only available for semilinear problems. krylov=true is much faster for larger systems and is thus recommended whenever there are >100 ODEs.

m: Controls the size of Krylov subspace.

iop: If not zero, determines the length of the incomplete orthogonalization procedure (IOP). Note that if the linear operator/Jacobian is hermitian, then the Lanczos algorithm will always be used and the IOP setting is ignored.

References

Hochbruck, Marlis, and Alexander Ostermann. “Exponential Integrators.” Acta Numerica 19 (2010): 209–286. doi:10.1017/S0962492910000048.

OrdinaryDiffEqExponentialRK.ETDRK3 — Type

ETDRK3(; krylov = false,
         m = 30,
         iop = 0)

Semilinear ODE solver 3rd order exponential-RK scheme.

Keyword Arguments

krylov: Determines whether Krylov approximation or operator caching is used, the latter only available for semilinear problems. krylov=true is much faster for larger systems and is thus recommended whenever there are >100 ODEs.

m: Controls the size of Krylov subspace.

iop: If not zero, determines the length of the incomplete orthogonalization procedure (IOP). Note that if the linear operator/Jacobian is hermitian, then the Lanczos algorithm will always be used and the IOP setting is ignored.

References

Hochbruck, Marlis, and Alexander Ostermann. “Exponential Integrators.” Acta Numerica 19 (2010): 209–286. doi:10.1017/S0962492910000048.

OrdinaryDiffEqExponentialRK.ETDRK4 — Type

ETDRK4(; krylov = false,
         m = 30,
         iop = 0)

Semilinear ODE solver 4th order exponential-RK scheme

Keyword Arguments

krylov: Determines whether Krylov approximation or operator caching is used, the latter only available for semilinear problems. krylov=true is much faster for larger systems and is thus recommended whenever there are >100 ODEs.

m: Controls the size of Krylov subspace.

iop: If not zero, determines the length of the incomplete orthogonalization procedure (IOP). Note that if the linear operator/Jacobian is hermitian, then the Lanczos algorithm will always be used and the IOP setting is ignored.

References

Hochbruck, Marlis, and Alexander Ostermann. “Exponential Integrators.” Acta Numerica 19 (2010): 209–286. doi:10.1017/S0962492910000048.

OrdinaryDiffEqExponentialRK.HochOst4 — Type

HochOst4(; krylov = false,
           m = 30,
           iop = 0)

Semilinear ODE solver 4th order exponential-RK scheme with stiff order 4.

Keyword Arguments

krylov: Determines whether Krylov approximation or operator caching is used, the latter only available for semilinear problems. krylov=true is much faster for larger systems and is thus recommended whenever there are >100 ODEs.

m: Controls the size of Krylov subspace.

iop: If not zero, determines the length of the incomplete orthogonalization procedure (IOP). Note that if the linear operator/Jacobian is hermitian, then the Lanczos algorithm will always be used and the IOP setting is ignored.

References

Hochbruck, Marlis, and Alexander Ostermann. “Exponential Integrators.” Acta Numerica 19 (2010): 209–286. doi:10.1017/S0962492910000048.

Adaptive Exponential Rosenbrock Methods

OrdinaryDiffEqExponentialRK.Exprb32 — Type

Exprb32(; m = 30,
          iop = 0)

Semilinear ODE solver 3rd order adaptive Exponential-Rosenbrock scheme.

Keyword Arguments

m: Controls the size of Krylov subspace.

iop: If not zero, determines the length of the incomplete orthogonalization procedure (IOP). Note that if the linear operator/Jacobian is hermitian, then the Lanczos algorithm will always be used and the IOP setting is ignored.

References

Hochbruck, M., & Ostermann, A. (2010). Exponential integrators. Acta Numerica, 19, 209-286. (https://doi.org/10.1017/S0962492910000048)

OrdinaryDiffEqExponentialRK.Exprb43 — Type

Exprb43(; m = 30,
          iop = 0)

Semilinear ODE solver 4th order adaptive Exponential-Rosenbrock scheme.

Keyword Arguments

m: Controls the size of Krylov subspace.

iop: If not zero, determines the length of the incomplete orthogonalization procedure (IOP). Note that if the linear operator/Jacobian is hermitian, then the Lanczos algorithm will always be used and the IOP setting is ignored.

References

Hochbruck, M., & Ostermann, A. (2010). Exponential integrators. Acta Numerica, 19, 209-286. (https://doi.org/10.1017/S0962492910000048)

Exponential Propagation Iterative Runge-Kutta Methods (EPIRK)

Missing docstring.

Missing docstring for Exp4. Check Documenter's build log for details.

Missing docstring.

Missing docstring for EPIRK4s3A. Check Documenter's build log for details.

Missing docstring.

Missing docstring for EPIRK4s3B. Check Documenter's build log for details.

Missing docstring.

Missing docstring for EPIRK5s3. Check Documenter's build log for details.

Missing docstring.

Missing docstring for EXPRB53s3. Check Documenter's build log for details.

Missing docstring.

Missing docstring for EPIRK5P1. Check Documenter's build log for details.

Missing docstring.

Missing docstring for EPIRK5P2. Check Documenter's build log for details.