Implicit Extrapolation Methods

OrdinaryDiffEq.OrdinaryDiffEqExtrapolation.ImplicitEulerExtrapolation — Type

ImplicitEulerExtrapolation: Parallelized Implicit Extrapolation Method Extrapolation of implicit Euler method with Romberg sequence. Similar to Hairer's SEULEX.

OrdinaryDiffEq.OrdinaryDiffEqExtrapolation.ImplicitDeuflhardExtrapolation — Type

ImplicitDeuflhardExtrapolation: Parallelized Implicit Extrapolation Method Midpoint extrapolation using Barycentric coordinates

OrdinaryDiffEq.OrdinaryDiffEqExtrapolation.ImplicitHairerWannerExtrapolation — Type

ImplicitHairerWannerExtrapolation: Parallelized Implicit Extrapolation Method Midpoint extrapolation using Barycentric coordinates, following Hairer's SODEX in the adaptivity behavior.

OrdinaryDiffEq.OrdinaryDiffEqExtrapolation.ImplicitEulerBarycentricExtrapolation — Type

ImplicitEulerBarycentricExtrapolation: Parallelized Implicit Extrapolation Method Euler extrapolation using Barycentric coordinates, following Hairer's SODEX in the adaptivity behavior.