Rosenbrock Methods

Standard Rosenbrock Methods

OrdinaryDiffEq.ROS2 — Type

ROS2()

2nd order stiffly accurate Rosenbrock-Wanner method with 2 internal stages with (Rinf=0). The embedded method is taken from Kinetic PreProcessor (KPP). J. G. Verwer et al. (1999): A second-order Rosenbrock method applied to photochemical dispersion problems More Information at https://doi.org/10.1137/S1064827597326651

OrdinaryDiffEq.ROS3 — Type

3rd order L-stable Rosenbrock-Wanner method with 3 internal stages with an embedded strongly A-stable 2nd order method More Information at https://link.springer.com/book/10.1007/978-3-662-09947-6

OrdinaryDiffEq.ROS2PR — Type

2nd order stiffly accurate Rosenbrock-Wanner method with 3 internal stages with (Rinf=0). For problems with medium stiffness the convergence behaviour is very poor and it is recommended to use ROS2S instead. More Information at https://doi.org/10.24355/dbbs.084-201408121139-0

OrdinaryDiffEq.ROS3PR — Type

3nd order stiffly accurate Rosenbrock-Wanner method with 3 internal stages with B_PR consistent of order 3, which is strongly A-stable with Rinf~=-0.73. More Information at https://doi.org/10.24355/dbbs.084-201408121139-0

Missing docstring.

Missing docstring for Scholz47. Check Documenter's build log for details.

OrdinaryDiffEq.ROS3PRL — Type

3rd order stiffly accurate Rosenbrock-Wanner method with 4 internal stages, B_PR consistent of order 2 with Rinf=0. The order of convergence decreases if medium stiff problems are considered, but it has good results for very stiff cases. More Information at https://doi.org/10.24355/dbbs.084-201408121139-0

OrdinaryDiffEq.ROS3PRL2 — Type

3rd order stiffly accurate Rosenbrock-Wanner method with 4 internal stages, B_PR consistent of order 3. The order of convergence does NOT decreases if medium stiff problems are considered as it does for ROS3PRL. More Information at https://doi.org/10.24355/dbbs.084-201408121139-0

Missing docstring.

Missing docstring for ROS3P. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Rodas3. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Rodas3P. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RosShamp4. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Veldd4. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Velds4. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GRK4T. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GRK4A. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Ros4LStab. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Rodas4. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Rodas42. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Rodas4P. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Rodas4P2. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Rodas5. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Rodas5P. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GeneralRosenbrock. Check Documenter's build log for details.

Rosenbrock W-Methods

Missing docstring.

Missing docstring for Rosenbrock23. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Rosenbrock32. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Rodas23W. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ROS34PW1a. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ROS34PW1b. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ROS34PW2. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ROS34PW3. Check Documenter's build log for details.

OrdinaryDiffEq.ROS34PRw — Type

3rd order stiffly accurate Rosenbrock-Wanner W-method with 4 internal stages, B_PR consistent of order 2. The order of convergence decreases if medium stiff problems are considered. More Information at https://doi.org/10.1016/j.cam.2015.03.010

OrdinaryDiffEq.ROS2S — Type

2nd order stiffly accurate Rosenbrock-Wanner W-method with 3 internal stages with B_PR consistent of order 2 with (Rinf=0). More Information at https://doi.org/10.24355/dbbs.084-201408121139-0

OrdinaryDiffEq.RosenbrockW6S4OS — Type

RosenbrockW6S4OS: Rosenbrock-W Method A 4th order L-stable Rosenbrock-W method (fixed step only).