Multistep Methods for Stiff Equations

OrdinaryDiffEq.OrdinaryDiffEqBDF.QNDF1 — Type

QNDF1: Multistep Method An adaptive order 1 quasi-constant timestep L-stable numerical differentiation function (NDF) method. Optional parameter kappa defaults to Shampine's accuracy-optimal -0.1850.

See also QNDF.

OrdinaryDiffEq.OrdinaryDiffEqBDF.QBDF1 — Function

QBDF1: Multistep Method

An alias of QNDF1 with κ=0.

OrdinaryDiffEq.OrdinaryDiffEqBDF.QNDF2 — Type

QNDF2: Multistep Method An adaptive order 2 quasi-constant timestep L-stable numerical differentiation function (NDF) method.

See also QNDF.

OrdinaryDiffEq.OrdinaryDiffEqBDF.QBDF2 — Function

QBDF2: Multistep Method

An alias of QNDF2 with κ=0.

OrdinaryDiffEq.OrdinaryDiffEqBDF.ABDF2 — Type

E. Alberdi Celayaa, J. J. Anza Aguirrezabalab, P. Chatzipantelidisc. Implementation of an Adaptive BDF2 Formula and Comparison with The MATLAB Ode15s. Procedia Computer Science, 29, pp 1014-1026, 2014. doi: https://doi.org/10.1016/j.procs.2014.05.091

ABDF2: Multistep Method An adaptive order 2 L-stable fixed leading coefficient multistep BDF method.

OrdinaryDiffEq.OrdinaryDiffEqBDF.QNDF — Type

QNDF: Multistep Method An adaptive order quasi-constant timestep NDF method. Utilizes Shampine's accuracy-optimal kappa values as defaults (has a keyword argument for a tuple of kappa coefficients).

@article{shampine1997matlab, title={The matlab ode suite}, author={Shampine, Lawrence F and Reichelt, Mark W}, journal={SIAM journal on scientific computing}, volume={18}, number={1}, pages={1–22}, year={1997}, publisher={SIAM} }

OrdinaryDiffEq.OrdinaryDiffEqBDF.QBDF — Function

QBDF: Multistep Method

An alias of QNDF with κ=0.

OrdinaryDiffEq.OrdinaryDiffEqBDF.FBDF — Type

FBDF: Fixed leading coefficient BDF

An adaptive order quasi-constant timestep NDF method. Utilizes Shampine's accuracy-optimal kappa values as defaults (has a keyword argument for a tuple of kappa coefficients).

@article{shampine2002solving, title={Solving 0= F (t, y (t), y′(t)) in Matlab}, author={Shampine, Lawrence F}, year={2002}, publisher={Walter de Gruyter GmbH \& Co. KG} }

OrdinaryDiffEq.OrdinaryDiffEqBDF.MEBDF2 — Type

MEBDF2: Multistep Method The second order Modified Extended BDF method, which has improved stability properties over the standard BDF. Fixed timestep only.