Rosenbrock Methods

Standard Rosenbrock Methods

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.ROS2 — Type

ROS2(; autodiff = Val{true}(),
       concrete_jac = nothing,
       linsolve = nothing,
       precs = DEFAULT_PRECS)

Rosenbrock Method. A 2nd order L-stable Rosenbrock method with 2 internal stages.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

References

J. G. Verwer et al. (1999): A second-order Rosenbrock method applied to photochemical dispersion problems https://doi.org/10.1137/S1064827597326651

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.ROS3 — Type

ROS3(; autodiff = Val{true}(),
       concrete_jac = nothing,
       linsolve = nothing,
       precs = DEFAULT_PRECS)

Rosenbrock Method. 3rd order L-stable Rosenbrock method with 3 internal stages with an embedded strongly A-stable 2nd order method.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

References

E. Hairer, G. Wanner, Solving ordinary differential equations II, stiff and differential-algebraic problems. Computational mathematics (2nd revised ed.), Springer (1996)

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.ROS2PR — Type

ROS2PR(; autodiff = Val{true}(),
         concrete_jac = nothing,
         linsolve = nothing,
         precs = DEFAULT_PRECS)

Rosenbrock Method. 2nd order stiffly accurate Rosenbrock method with 3 internal stages with (Rinf=0). For problems with medium stiffness the convergence behaviour is very poor and it is recommended to use ROS2S instead.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

References

Rang, Joachim (2014): The Prothero and Robinson example: Convergence studies for Runge-Kutta and Rosenbrock-Wanner methods. https://doi.org/10.24355/dbbs.084-201408121139-0

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.ROS3PR — Type

ROS3PR(; autodiff = Val{true}(),
         concrete_jac = nothing,
         linsolve = nothing,
         precs = DEFAULT_PRECS)

Rosenbrock Method. 3nd order stiffly accurate Rosenbrock method with 3 internal stages with B_PR consistent of order 3, which is strongly A-stable with Rinf~=-0.73.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

References

Rang, Joachim (2014): The Prothero and Robinson example: Convergence studies for Runge-Kutta and Rosenbrock-Wanner methods. https://doi.org/10.24355/dbbs.084-201408121139-0

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.Scholz4_7 — Type

Scholz4_7(; autodiff = Val{true}(),
            concrete_jac = nothing,
            linsolve = nothing,
            precs = DEFAULT_PRECS)

Rosenbrock Method. 3nd order stiffly accurate Rosenbrock method with 3 internal stages with B_PR consistent of order 3, which is strongly A-stable with Rinf~=-0.73. Convergence with order 4 for the stiff case, but has a poor accuracy.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

References

Rang, Joachim (2014): The Prothero and Robinson example: Convergence studies for Runge-Kutta and Rosenbrock-Wanner methods. https://doi.org/10.24355/dbbs.084-201408121139-0

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.ROS3PRL — Type

ROS3PRL(; autodiff = Val{true}(),
          concrete_jac = nothing,
          linsolve = nothing,
          precs = DEFAULT_PRECS)

Rosenbrock Method. 3rd order stiffly accurate Rosenbrock method with 4 internal stages, B_PR consistent of order 2 with Rinf=0. The order of convergence decreases if medium stiff problems are considered, but it has good results for very stiff cases.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

References

Rang, Joachim (2014): The Prothero and Robinson example: Convergence studies for Runge-Kutta and Rosenbrock-Wanner methods. https://doi.org/10.24355/dbbs.084-201408121139-0

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.ROS3PRL2 — Type

ROS3PRL2(; autodiff = Val{true}(),
           concrete_jac = nothing,
           linsolve = nothing,
           precs = DEFAULT_PRECS)

Rosenbrock Method. 3rd order stiffly accurate Rosenbrock method with 4 internal stages, B_PR consistent of order 3. The order of convergence does NOT decreases if medium stiff problems are considered as it does for ROS3PRL.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

References

Rang, Joachim (2014): The Prothero and Robinson example: Convergence studies for Runge-Kutta and Rosenbrock-Wanner methods. https://doi.org/10.24355/dbbs.084-201408121139-0

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.ROS3P — Type

ROS3P(; autodiff = Val{true}(),
        concrete_jac = nothing,
        linsolve = nothing,
        precs = DEFAULT_PRECS,
        step_limiter! = OrdinaryDiffEq.trivial_limiter!)

Rosenbrock Method. 3rd order A-stable and stiffly stable Rosenbrock method. Keeps high accuracy on discretizations of nonlinear parabolic PDEs.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

step_limiter!: function of the form limiter!(u, integrator, p, t)

References

Lang, J. & Verwer, ROS3P—An Accurate Third-Order Rosenbrock Solver Designed for Parabolic Problems J. BIT Numerical Mathematics (2001) 41: 731. doi:10.1023/A:1021900219772

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.Rodas3 — Type

Rodas3(; autodiff = Val{true}(),
         concrete_jac = nothing,
         linsolve = nothing,
         precs = DEFAULT_PRECS,
         step_limiter! = OrdinaryDiffEq.trivial_limiter!)

Rosenbrock Method. 3rd order A-stable and stiffly stable Rosenbrock method.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

step_limiter!: function of the form limiter!(u, integrator, p, t)

References

Steinebach G. Construction of Rosenbrock–Wanner method Rodas5P and numerical benchmarks within the Julia Differential Equations package. In: BIT Numerical Mathematics, 63(2), 2023

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.Rodas3P — Type

Rodas3P(; autodiff = Val{true}(),
          concrete_jac = nothing,
          linsolve = nothing,
          precs = DEFAULT_PRECS,
          step_limiter! = OrdinaryDiffEq.trivial_limiter!)

Rosenbrock Method. 3rd order A-stable and stiffly stable Rosenbrock method with a stiff-aware 3rd order interpolant and additional error test for interpolation. Keeps accuracy on discretizations of linear parabolic PDEs.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

step_limiter!: function of the form limiter!(u, integrator, p, t)

References

Steinebach G., Rodas23W / Rodas32P - a Rosenbrock-type method for DAEs with additional error estimate for dense output and Julia implementation, In progress.

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.RosShamp4 — Type

RosShamp4(; autodiff = Val{true}(),
            concrete_jac = nothing,
            linsolve = nothing,
            precs = DEFAULT_PRECS)

Rosenbrock Method. An A-stable 4th order Rosenbrock method.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

References

L. F. Shampine, Implementation of Rosenbrock Methods, ACM Transactions on Mathematical Software (TOMS), 8: 2, 93-113. doi:10.1145/355993.355994

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.Veldd4 — Type

Veldd4(; autodiff = Val{true}(),
         concrete_jac = nothing,
         linsolve = nothing,
         precs = DEFAULT_PRECS,
         step_limiter! = OrdinaryDiffEq.trivial_limiter!)

Rosenbrock Method. A 4th order D-stable Rosenbrock method.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

step_limiter!: function of the form limiter!(u, integrator, p, t)

References

van Veldhuizen, D-stability and Kaps-Rentrop-methods, M. Computing (1984) 32: 229. doi:10.1007/BF02243574

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.Velds4 — Type

Velds4(; autodiff = Val{true}(),
         concrete_jac = nothing,
         linsolve = nothing,
         precs = DEFAULT_PRECS,
         step_limiter! = OrdinaryDiffEq.trivial_limiter!)

Rosenbrock Method. A 4th order A-stable Rosenbrock method.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

step_limiter!: function of the form limiter!(u, integrator, p, t)

References

van Veldhuizen, D-stability and Kaps-Rentrop-methods, M. Computing (1984) 32: 229. doi:10.1007/BF02243574

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.GRK4T — Type

GRK4T(; autodiff = Val{true}(),
        concrete_jac = nothing,
        linsolve = nothing,
        precs = DEFAULT_PRECS,
        step_limiter! = OrdinaryDiffEq.trivial_limiter!)

Rosenbrock Method. An efficient 4th order Rosenbrock method.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

step_limiter!: function of the form limiter!(u, integrator, p, t)

References

Kaps, P. & Rentrop, Generalized Runge-Kutta methods of order four with stepsize control for stiff ordinary differential equations. P. Numer. Math. (1979) 33: 55. doi:10.1007/BF01396495

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.GRK4A — Type

GRK4A(; autodiff = Val{true}(),
        concrete_jac = nothing,
        linsolve = nothing,
        precs = DEFAULT_PRECS,
        step_limiter! = OrdinaryDiffEq.trivial_limiter!)

Rosenbrock Method. An A-stable 4th order Rosenbrock method. Essentially "anti-L-stable" but efficient.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

step_limiter!: function of the form limiter!(u, integrator, p, t)

References

Kaps, P. & Rentrop, Generalized Runge-Kutta methods of order four with stepsize control for stiff ordinary differential equations. P. Numer. Math. (1979) 33: 55. doi:10.1007/BF01396495

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.Ros4LStab — Type

Ros4LStab(; autodiff = Val{true}(),
            concrete_jac = nothing,
            linsolve = nothing,
            precs = DEFAULT_PRECS,
            step_limiter! = OrdinaryDiffEq.trivial_limiter!)

Rosenbrock Method. A 4th order A-stable stiffly stable Rosenbrock method with a stiff-aware 3rd order interpolant

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

step_limiter!: function of the form limiter!(u, integrator, p, t)

References

E. Hairer, G. Wanner, Solving ordinary differential equations II, stiff and differential-algebraic problems. Computational mathematics (2nd revised ed.), Springer (1996)

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.Rodas4 — Type

Rodas4(; autodiff = Val{true}(),
         concrete_jac = nothing,
         linsolve = nothing,
         precs = DEFAULT_PRECS,
         step_limiter! = OrdinaryDiffEq.trivial_limiter!)

Rosenbrock Method. A 4th order L-stable Rosenbrock method.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

step_limiter!: function of the form limiter!(u, integrator, p, t)

References

E. Hairer, G. Wanner, Solving ordinary differential equations II, stiff and differential-algebraic problems. Computational mathematics (2nd revised ed.), Springer (1996)

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.Rodas42 — Type

Rodas42(; autodiff = Val{true}(),
          concrete_jac = nothing,
          linsolve = nothing,
          precs = DEFAULT_PRECS,
          step_limiter! = OrdinaryDiffEq.trivial_limiter!)

Rosenbrock Method. A 4th order A-stable stiffly stable Rosenbrock method with a stiff-aware 3rd order interpolant

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

step_limiter!: function of the form limiter!(u, integrator, p, t)

References

E. Hairer, G. Wanner, Solving ordinary differential equations II, stiff and differential-algebraic problems. Computational mathematics (2nd revised ed.), Springer (1996)

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.Rodas4P — Type

Rodas4P(; autodiff = Val{true}(),
          concrete_jac = nothing,
          linsolve = nothing,
          precs = DEFAULT_PRECS,
          step_limiter! = OrdinaryDiffEq.trivial_limiter!)

Rosenbrock-Wanner Method. 4th order A-stable stiffly stable Rosenbrock method with a stiff-aware 3rd order interpolant. 4th order on linear parabolic problems and 3rd order accurate on nonlinear parabolic problems (as opposed to lower if not corrected).

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

step_limiter!: function of the form limiter!(u, integrator, p, t)

References

Steinebach G., Rodas23W / Rodas32P - a Rosenbrock-type method for DAEs with additional error estimate for dense output and Julia implementation, In progress.

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.Rodas4P2 — Type

Rodas4P2(; autodiff = Val{true}(),
           concrete_jac = nothing,
           linsolve = nothing,
           precs = DEFAULT_PRECS,
           step_limiter! = OrdinaryDiffEq.trivial_limiter!)

Rosenbrock-Wanner Method. A 4th order L-stable stiffly stable Rosenbrock method with a stiff-aware 3rd order interpolant. 4th order on linear parabolic problems and 3rd order accurate on nonlinear parabolic problems. It is an improvement of Roadas4P and in case of inexact Jacobians a second order W method.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

step_limiter!: function of the form limiter!(u, integrator, p, t)

References

Steinebach G., Rodas23W / Rodas32P - a Rosenbrock-type method for DAEs with additional error estimate for dense output and Julia implementation, In progress.

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.Rodas5 — Type

Rodas5(; autodiff = Val{true}(),
         concrete_jac = nothing,
         linsolve = nothing,
         precs = DEFAULT_PRECS,
         step_limiter! = OrdinaryDiffEq.trivial_limiter!)

Rosenbrock Method. A 5th order A-stable stiffly stable Rosenbrock method with a stiff-aware 4th order interpolant.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

step_limiter!: function of the form limiter!(u, integrator, p, t)

References

Di Marzo G. RODAS5(4) – Méthodes de Rosenbrock d’ordre 5(4) adaptées aux problemes différentiels-algébriques. MSc mathematics thesis, Faculty of Science, University of Geneva, Switzerland.

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.Rodas5P — Type

Rodas5P(; autodiff = Val{true}(),
          concrete_jac = nothing,
          linsolve = nothing,
          precs = DEFAULT_PRECS,
          step_limiter! = OrdinaryDiffEq.trivial_limiter!)

Rosenbrock Method. A 5th order A-stable stiffly stable Rosenbrock method with a stiff-aware 4th order interpolant. Has improved stability in the adaptive time stepping embedding.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

step_limiter!: function of the form limiter!(u, integrator, p, t)

References

Steinebach G. Construction of Rosenbrock–Wanner method Rodas5P and numerical benchmarks within the Julia Differential Equations package. In: BIT Numerical Mathematics, 63(2), 2023

Rosenbrock W-Methods

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.Rosenbrock23 — Type

Rosenbrock23(; autodiff = Val{true}(),
               concrete_jac = nothing,
               linsolve = nothing,
               precs = DEFAULT_PRECS,
               step_limiter! = OrdinaryDiffEq.trivial_limiter!)

Rosenbrock-Wanner Method. An Order 2/3 L-Stable Rosenbrock-W method which is good for very stiff equations with oscillations at low tolerances. 2nd order stiff-aware interpolation.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

step_limiter!: function of the form limiter!(u, integrator, p, t)

References

Shampine L.F. and Reichelt M., (1997) The MATLAB ODE Suite, SIAM Journal of

Scientific Computing, 18 (1), pp. 1-22.

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.Rosenbrock32 — Type

Rosenbrock32(; autodiff = Val{true}(),
               concrete_jac = nothing,
               linsolve = nothing,
               precs = DEFAULT_PRECS,
               step_limiter! = OrdinaryDiffEq.trivial_limiter!)

Rosenbrock-Wanner Method. An Order 3/2 A-Stable Rosenbrock-W method which is good for mildly stiff equations without oscillations at low tolerances. Note that this method is prone to instability in the presence of oscillations, so use with caution. 2nd order stiff-aware interpolation.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

step_limiter!: function of the form limiter!(u, integrator, p, t)

References

Shampine L.F. and Reichelt M., (1997) The MATLAB ODE Suite, SIAM Journal of

Scientific Computing, 18 (1), pp. 1-22.

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.Rodas23W — Type

Rodas23W(; autodiff = Val{true}(),
           concrete_jac = nothing,
           linsolve = nothing,
           precs = DEFAULT_PRECS,
           step_limiter! = OrdinaryDiffEq.trivial_limiter!)

Rosenbrock-Wanner Method. An Order 2/3 L-Stable Rosenbrock-W method for stiff ODEs and DAEs in mass matrix form. 2nd order stiff-aware interpolation and additional error test for interpolation.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

step_limiter!: function of the form limiter!(u, integrator, p, t)

References

Steinebach G., Rodas23W / Rodas32P - a Rosenbrock-type method for DAEs with additional error estimate for dense output and Julia implementation, In progress.

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.ROS34PW1a — Type

ROS34PW1a(; autodiff = Val{true}(),
            concrete_jac = nothing,
            linsolve = nothing,
            precs = DEFAULT_PRECS)

Rosenbrock-Wanner Method. A 4th order L-stable Rosenbrock-W method.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

References

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.ROS34PW1b — Type

ROS34PW1b(; autodiff = Val{true}(),
            concrete_jac = nothing,
            linsolve = nothing,
            precs = DEFAULT_PRECS)

Rosenbrock-Wanner Method. A 4th order L-stable Rosenbrock-W method.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

References

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.ROS34PW2 — Type

ROS34PW2(; autodiff = Val{true}(),
           concrete_jac = nothing,
           linsolve = nothing,
           precs = DEFAULT_PRECS)

Rosenbrock-Wanner Method. A 4th order stiffy accurate Rosenbrock-W method for PDAEs.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

References

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.ROS34PW3 — Type

ROS34PW3(; autodiff = Val{true}(),
           concrete_jac = nothing,
           linsolve = nothing,
           precs = DEFAULT_PRECS)

Rosenbrock-Wanner Method. A 4th order strongly A-stable (Rinf~0.63) Rosenbrock-W method.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

References

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.ROS34PRw — Type

ROS34PRw(; autodiff = Val{true}(),
           concrete_jac = nothing,
           linsolve = nothing,
           precs = DEFAULT_PRECS)

Rosenbrock-Wanner Method. 3rd order stiffly accurate Rosenbrock-Wanner W-method with 4 internal stages, B_PR consistent of order 2. The order of convergence decreases if medium stiff problems are considered.

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

References

Joachim Rang, Improved traditional Rosenbrock–Wanner methods for stiff ODEs and DAEs, Journal of Computational and Applied Mathematics, https://doi.org/10.1016/j.cam.2015.03.010

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.ROK4a — Type

ROK4a(; autodiff = Val{true}(),
        concrete_jac = nothing,
        linsolve = nothing,
        precs = DEFAULT_PRECS)

Rosenbrock-Wanner Method. 4rd order L-stable Rosenbrock-Krylov method with 4 internal stages, with a 3rd order embedded method which is strongly A-stable with Rinf~=0.55. (when using exact Jacobians)

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

References

Tranquilli, Paul and Sandu, Adrian (2014): Rosenbrock–Krylov Methods for Large Systems of Differential Equations https://doi.org/10.1137/130923336

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.ROS2S — Type

ROS2S(; autodiff = Val{true}(),
        concrete_jac = nothing,
        linsolve = nothing,
        precs = DEFAULT_PRECS)

Rosenbrock-Wanner Method. 2nd order stiffly accurate Rosenbrock-Wanner W-method with 3 internal stages with B_PR consistent of order 2 with (Rinf=0).

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

References

Rang, Joachim (2014): The Prothero and Robinson example: Convergence studies for Runge-Kutta and Rosenbrock-Wanner methods. https://doi.org/10.24355/dbbs.084-201408121139-0

OrdinaryDiffEq.OrdinaryDiffEqRosenbrock.RosenbrockW6S4OS — Type

RosenbrockW6S4OS(; autodiff = Val{true}(),
                   concrete_jac = nothing,
                   linsolve = nothing,
                   precs = DEFAULT_PRECS)

Rosenbrock-Wanner Method. A 4th order L-stable Rosenbrock-W method (fixed step only).

Keyword Arguments

autodiff: boolean to control if the Jacobian should be computed via AD or not

concrete_jac: function of the form jac!(J, u, p, t)

linsolve: custom solver for the inner linear systems

precs: custom preconditioner for the inner linear solver

References

https://doi.org/10.1016/j.cam.2009.09.017