Singly-Diagonally Implicit Runge-Kutta (SDIRK) Methods

OrdinaryDiffEq.ImplicitEuler — Type

ImplicitEuler: SDIRK Method A 1st order implicit solver. A-B-L-stable. Adaptive timestepping through a divided differences estimate via memory. Strong-stability preserving (SSP).

OrdinaryDiffEq.ImplicitMidpoint — Type

ImplicitMidpoint: SDIRK Method A second order A-stable symplectic and symmetric implicit solver. Good for highly stiff equations which need symplectic integration.

OrdinaryDiffEq.Trapezoid — Type

Andre Vladimirescu. 1994. The Spice Book. John Wiley & Sons, Inc., New York, NY, USA.

Trapezoid: SDIRK Method A second order A-stable symmetric ESDIRK method. "Almost symplectic" without numerical dampening. Also known as Crank-Nicolson when applied to PDEs. Adaptive timestepping via divided

OrdinaryDiffEq.TRBDF2 — Type

@article{hosea1996analysis, title={Analysis and implementation of TR-BDF2}, author={Hosea, ME and Shampine, LF}, journal={Applied Numerical Mathematics}, volume={20}, number={1-2}, pages={21–37}, year={1996}, publisher={Elsevier} }

TRBDF2: SDIRK Method A second order A-B-L-S-stable one-step ESDIRK method. Includes stiffness-robust error estimates for accurate adaptive timestepping, smoothed derivatives for highly stiff and oscillatory problems.

OrdinaryDiffEq.SDIRK2 — Type

@article{hindmarsh2005sundials, title={{SUNDIALS}: Suite of nonlinear and differential/algebraic equation solvers}, author={Hindmarsh, Alan C and Brown, Peter N and Grant, Keith E and Lee, Steven L and Serban, Radu and Shumaker, Dan E and Woodward, Carol S}, journal={ACM Transactions on Mathematical Software (TOMS)}, volume={31}, number={3}, pages={363–396}, year={2005}, publisher={ACM} }

SDIRK2: SDIRK Method An A-B-L stable 2nd order SDIRK method

Missing docstring.

Missing docstring for SDIRK22. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SSPSDIRK2. Check Documenter's build log for details.

OrdinaryDiffEq.Kvaerno3 — Type

@article{kvaerno2004singly, title={Singly diagonally implicit Runge–Kutta methods with an explicit first stage}, author={Kv{\ae}rn{\o}, Anne}, journal={BIT Numerical Mathematics}, volume={44}, number={3}, pages={489–502}, year={2004}, publisher={Springer} }

Kvaerno3: SDIRK Method An A-L stable stiffly-accurate 3rd order ESDIRK method

Missing docstring.

Missing docstring for CFNLIRK3. Check Documenter's build log for details.

OrdinaryDiffEq.Cash4 — Type

@article{hindmarsh2005sundials, title={{SUNDIALS}: Suite of nonlinear and differential/algebraic equation solvers}, author={Hindmarsh, Alan C and Brown, Peter N and Grant, Keith E and Lee, Steven L and Serban, Radu and Shumaker, Dan E and Woodward, Carol S}, journal={ACM Transactions on Mathematical Software (TOMS)}, volume={31}, number={3}, pages={363–396}, year={2005}, publisher={ACM} }

Cash4: SDIRK Method An A-L stable 4th order SDIRK method

Missing docstring.

Missing docstring for SFSDIRK4. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SFSDIRK5. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SFSDIRK6. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SFSDIRK7. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SFSDIRK8. Check Documenter's build log for details.

OrdinaryDiffEq.Hairer4 — Type

E. Hairer, G. Wanner, Solving ordinary differential equations II, stiff and differential-algebraic problems. Computational mathematics (2nd revised ed.), Springer (1996)

Hairer4: SDIRK Method An A-L stable 4th order SDIRK method

OrdinaryDiffEq.Hairer42 — Type

E. Hairer, G. Wanner, Solving ordinary differential equations II, stiff and differential-algebraic problems. Computational mathematics (2nd revised ed.), Springer (1996)

Hairer42: SDIRK Method An A-L stable 4th order SDIRK method

OrdinaryDiffEq.Kvaerno4 — Type

@article{kvaerno2004singly, title={Singly diagonally implicit Runge–Kutta methods with an explicit first stage}, author={Kv{\ae}rn{\o}, Anne}, journal={BIT Numerical Mathematics}, volume={44}, number={3}, pages={489–502}, year={2004}, publisher={Springer} }

Kvaerno4: SDIRK Method An A-L stable stiffly-accurate 4th order ESDIRK metho

OrdinaryDiffEq.Kvaerno5 — Type

@article{kvaerno2004singly, title={Singly diagonally implicit Runge–Kutta methods with an explicit first stage}, author={Kv{\ae}rn{\o}, Anne}, journal={BIT Numerical Mathematics}, volume={44}, number={3}, pages={489–502}, year={2004}, publisher={Springer} }

Kvaerno5: SDIRK Method An A-L stable stiffly-accurate 5th order ESDIRK method