Stabilized Runge-Kutta Methods (Runge-Kutta-Chebyshev)

Explicit Stabilized Runge-Kutta Methods

OrdinaryDiffEq.ROCK2 — Type

Assyr Abdulle, Alexei A. Medovikov. Second Order Chebyshev Methods based on Orthogonal Polynomials. Numerische Mathematik, 90 (1), pp 1-18, 2001. doi: https://dx.doi.org/10.1007/s002110100292

ROCK2: Stabilized Explicit Method Second order stabilized Runge-Kutta method. Exhibits high stability for real eigenvalues and is smoothened to allow for moderate sized complex eigenvalues.

OrdinaryDiffEq.ROCK4 — Type

Assyr Abdulle. Fourth Order Chebyshev Methods With Recurrence Relation. 2002 Society for Industrial and Applied Mathematics Journal on Scientific Computing, 23(6), pp 2041-2054, 2001. doi: https://doi.org/10.1137/S1064827500379549

ROCK4: Stabilized Explicit Method Fourth order stabilized Runge-Kutta method. Exhibits high stability for real eigenvalues and is smoothened to allow for moderate sized complex eigenvalues.

Missing docstring.

Missing docstring for ESERK4. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ESERK5. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RKC. Check Documenter's build log for details.

Implicit Stabilized Runge-Kutta Methods

Missing docstring.

Missing docstring for IRKC. Check Documenter's build log for details.